// src/app/shared/models/company.model.ts
export class CompanyModel {
    id: number;
    name: string;
    address: string;
    phone: string;
    business_no: string;
    gst_no: string;
    website: string;
    email: string;
    logo_link: string;
    user_id: number;
    first_name: string;
    last_name: string;
    created: Date;
    modified: Date;
}